implemented all present stubs and split tests for hashmap and btree in separate files
